refactor(types): типы из комментариев вынесены из инференсера
Третья группа знаний, не относящихся к выводу типов: тип, объявленный автором в
комментарии рядом с кодом. Два места, где комментарий признаётся объявлением, —
висячий комментарий объявления (`Перем Х; // Тип - Строка`) и висячий комментарий строки
присваивания (`Х = Ф(); // Тип - Строка`, встроенная типизация локальной переменной из
стандарта 1С:EDT).

Вынесено в `CommentDeclaredTypes`. Обратного вызова к выводу типов не нужно и здесь:
текст разбирает парсер описаний, а компонент только резолвит имена в реестре.

Инференсер: 1415 строк → 1350, от 1650 на develop.
